Improved cheat copying and select or fill operation.
I'm guessing a bit as to the plugin author's intentions, but for my use
case I find the functions to copy the cheat and more tellingly the
function to `select_or_fill_commandline` to be less than ideal. It seems
the author intended the former to copy the cheat code for easy
subsequent use but was copying only to the "0" register, which seems to
copy only to the " register and not more clipboard friendly registers.
I've changed this to include * as well. And in both cases I've changed
the author's use of the `vim.api.nvim_echo` method to add to message
history because the text doesn't appear anywhere at all if I don't. At
least that way I can review it in the message history.